Add comments to commands
